Introduce a lsp-fs protocol abstraction.
