+++ src/compiler/lsp/fs_connection_socket_posix.cc
@@ -13,7 +13,7 @@
 // The license can be found in the file `LICENSE` in the top level
 // directory of this repository.
 
-#include "../top.h"
+#include "../../top.h"
 
 #ifdef TOIT_POSIX
 
@@ -28,16 +28,15 @@
 #include <unistd.h>
 #include <string.h>
 
-#include "diagnostic.h"
-#include "filesystem_socket.h"
-#include "../utils.h"
+#include "fs_connection_socket.h"
+
+#include "../diagnostic.h"
+#include "../../utils.h"
 
 namespace toit {
 namespace compiler {
 
-char* get_executable_path();
-
-void FilesystemSocket::initialize(Diagnostics* diagnostics) {
+void LspFsConnectionSocket::initialize(Diagnostics* diagnostics) {
   if (_is_initialized) return;
   _is_initialized = true;
   addrinfo hints;
@@ -67,7 +66,7 @@ void FilesystemSocket::initialize(Diagnostics* diagnostics) {
   }
 }
 
-FilesystemSocket::~FilesystemSocket() {
+LspFsConnectionSocket::~LspFsConnectionSocket() {
   if (_socket != -1) {
     close(_socket);
     _socket = -1;
